Fix for the possibility of marking the wrong Timeout complete
There are complications using the default `setTimeout`-provided id when
referencing the Timeout to mark complete, since we don't get the id from
the default `setTimeout` until after we've committed to a handler. To
work around the issue, an app-specific random UUID is set for each
Timeout object, which is used in the handler to look up which Timeout to
mark as complete.

Refs #3, #4